move fuse driver
